+++ src/DroidSolutions.Oss.SemanticVersion/SemanticVersionObject.cs
@@ -1,5 +1,7 @@
 using System.Collections;
+using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
+using System.Reflection;
 using System.Text.RegularExpressions;
 
 namespace DroidSolutions.Oss.SemanticVersion;
@@ -7,7 +9,7 @@ namespace DroidSolutions.Oss.SemanticVersion;
 /// <summary>
 /// A model that represents a semantic version number.
 /// </summary>
-public class SemanticVersionObject : ISemanticVersion, IComparable
+public class SemanticVersionObject : ISemanticVersion, IComparable, IEquatable<SemanticVersionObject>
 {
   /// <summary>
   /// Initializes a new instance of the <see cref="SemanticVersionObject"/> class.
@@ -59,6 +61,23 @@ public class SemanticVersionObject : ISemanticVersion, IComparable
   }
 
   /// <summary>
+  /// Initializes a new instance of the <see cref="SemanticVersionObject"/> class.
+  /// </summary>
+  /// <param name="major">The major version.</param>
+  /// <param name="minor">The minor version.</param>
+  /// <param name="patch">The patch version.</param>
+  /// <param name="prerelease">The prerelease version.</param>
+  /// <param name="build">The build metadata.</param>
+  public SemanticVersionObject(int major, int minor, int patch, string? prerelease, string build)
+  {
+    Major = major;
+    Minor = minor;
+    Patch = patch;
+    PreRelease = prerelease;
+    Build = build;
+  }
+
+  /// <summary>
   /// Gets or sets the major version number.
   /// </summary>
   public virtual int Major { get; set; }
@@ -79,6 +98,77 @@ public class SemanticVersionObject : ISemanticVersion, IComparable
   public virtual string? PreRelease { get; set; }
 
   /// <summary>
+  /// Gets or sets the build metadata.
+  /// </summary>
+  public virtual string? Build { get; set; }
+
+  /// <summary>
+  /// Checks if both given items are equal.
+  /// </summary>
+  /// <param name="left">The left side of the comparison.</param>
+  /// <param name="right">The right side of the comparison.</param>
+  /// <returns><see langword="true"/> if both sides are equal, else <see langword="false"/>.</returns>
+  public static bool operator ==(SemanticVersionObject? left, SemanticVersionObject? right)
+  {
+    return left?.CompareTo(right) == 0;
+  }
+
+  /// <summary>
+  /// Checks if both given items are not equal.
+  /// </summary>
+  /// <param name="left">The left side of the comparison.</param>
+  /// <param name="right">The right side of the comparison.</param>
+  /// <returns><see langword="false"/> if both sides are equal, else <see langword="true"/>.</returns>
+  public static bool operator !=(SemanticVersionObject? left, SemanticVersionObject? right)
+  {
+    return left?.CompareTo(right) != 0;
+  }
+
+  /// <summary>
+  /// Checks if both given left is greater than right.
+  /// </summary>
+  /// <param name="left">The left side of the comparison.</param>
+  /// <param name="right">The right side of the comparison.</param>
+  /// <returns><see langword="true"/> if left is greater than the right side, else <see langword="false"/>.</returns>
+  public static bool operator >(SemanticVersionObject? left, SemanticVersionObject? right)
+  {
+    return left?.CompareTo(right) < 0;
+  }
+
+  /// <summary>
+  /// Checks if both given right is greater than left.
+  /// </summary>
+  /// <param name="left">The left side of the comparison.</param>
+  /// <param name="right">The right side of the comparison.</param>
+  /// <returns><see langword="true"/> if right is greater than the left side, else <see langword="false"/>.</returns>
+  public static bool operator <(SemanticVersionObject? left, SemanticVersionObject? right)
+  {
+    return left?.CompareTo(right) > 0;
+  }
+
+  /// <summary>
+  /// Checks if both given left is equal or greater than the right.
+  /// </summary>
+  /// <param name="left">The left side of the comparison.</param>
+  /// <param name="right">The right side of the comparison.</param>
+  /// <returns><see langword="true"/> if left is greater than the right side, else <see langword="false"/>.</returns>
+  public static bool operator >=(SemanticVersionObject? left, SemanticVersionObject? right)
+  {
+    return left?.CompareTo(right) <= 0;
+  }
+
+  /// <summary>
+  /// Checks if both given right is equal or greater than the left.
+  /// </summary>
+  /// <param name="left">The left side of the comparison.</param>
+  /// <param name="right">The right side of the comparison.</param>
+  /// <returns><see langword="true"/> if right is greater than the left side, else <see langword="false"/>.</returns>
+  public static bool operator <=(SemanticVersionObject? left, SemanticVersionObject? right)
+  {
+    return left?.CompareTo(right) >= 0;
+  }
+
+  /// <summary>
   /// Creates a <see cref="SemanticVersion"/> instance by parsing the given string.
   /// </summary>
   /// <param name="version">A version string in semantic release format, where the elements are separated by dots. The version can
@@ -86,7 +176,11 @@ public class SemanticVersionObject : ISemanticVersion, IComparable
   /// <returns>A <see cref="SemanticVersionObject"/> instance with the parsed version numbers. </returns>
   public static SemanticVersionObject FromString(string version)
   {
-    var regex = new Regex(@"^v?(\d+).(\d+).(\d+)(-(.*))?$", RegexOptions.CultureInvariant);
+    // @"^v?(\d+).(\d+).(\d+)(-(.*))?$",
+    Regex regex = new(
+      @"^v?(?<major>0|[1-9]\d*)\.(?<minor>0|[1-9]\d*)\.(?<patch>0|[1-9]\d*)(?:-(?<prerelease>(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\.(?:0|[1-9]\d*|\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\+(?<build>[0-9a-zA-Z-]+(?:\.[0-9a-zA-Z-]+)*))?$",
+      RegexOptions.CultureInvariant);
+
     Match match = regex.Match(version);
     if (!match.Success)
     {
@@ -94,19 +188,82 @@ public class SemanticVersionObject : ISemanticVersion, IComparable
         $"The given string \"{version}\" is not a valid semantic version!", nameof(version));
     }
 
-    string? preRelease = null;
-    if (match.Groups.Count > 5 && !string.IsNullOrWhiteSpace(match.Groups[5].Value))
+    SemanticVersionObject result = new(
+      int.Parse(match.Groups[1].Value, CultureInfo.InvariantCulture),
+      int.Parse(match.Groups[2].Value, CultureInfo.InvariantCulture),
+      int.Parse(match.Groups[3].Value, CultureInfo.InvariantCulture));
+
+    if (match.Groups.ContainsKey("prerelease"))
     {
-      preRelease = match.Groups[5].Value;
+      result.PreRelease = match.Groups["prerelease"].Value;
     }
 
-    return new SemanticVersionObject
+    if (match.Groups.ContainsKey("build"))
     {
-      Major = int.Parse(match.Groups[1].Value, CultureInfo.InvariantCulture),
-      Minor = int.Parse(match.Groups[2].Value, CultureInfo.InvariantCulture),
-      Patch = int.Parse(match.Groups[3].Value, CultureInfo.InvariantCulture),
-      PreRelease = preRelease,
-    };
+      result.Build = match.Groups["build"].Value;
+    }
+
+    return result;
+  }
+
+  /// <summary>
+  /// Creates a <see cref="SemanticVersion"/> instance by parsing the given <see cref="Version"/>.
+  /// </summary>
+  /// <remarks>
+  /// <para>The given <see cref="Version"/> object's properties are used to create an instance of the
+  /// <see cref="SemanticVersion"/> class. Major and minor properties are used accordingly.
+  /// </para>
+  /// <para>
+  /// <seealso href="https://learn.microsoft.com/en-us/dotnet/fundamentals/runtime-libraries/system-version">The docs
+  /// </seealso> state that revision should be used as patch version, however .NET AssemblyInfo.cs files are generated
+  /// with patch set to the build property. To control this, the <paramref name="useBuildAsPatch"/> parameter can be
+  /// used.
+  /// </para>
+  /// <para>
+  /// Note that Version objects don't support prereleases or build metadata as the build and revision properties are
+  /// ints.
+  /// </para>
+  /// </remarks>
+  /// <param name="version">The version to parse.</param>
+  /// <param name="useBuildAsPatch">If <see langword="true"/>, the <see cref="Version.Build"/> property will be
+  /// used as the patch version. Otherwise, the <see cref="Version.Revision"/> property will be used.</param>
+  /// <returns>A <see cref="SemanticVersionObject"/> instance with the parsed version numbers. </returns>
+  public static SemanticVersionObject FromVersion(Version version, bool useBuildAsPatch = false)
+  {
+    ArgumentNullException.ThrowIfNull(version);
+    int patch = useBuildAsPatch ? version.Build : version.Revision;
+    string? build = null;
+    if (useBuildAsPatch && version.Revision >= 1)
+    {
+      build = version.Revision.ToString();
+    }
+    else if (!useBuildAsPatch && version.Build >= 1)
+    {
+      build = version.Build.ToString();
+    }
+
+    return build is null
+      ? new SemanticVersionObject(version.Major, version.Minor, patch)
+      : new SemanticVersionObject(version.Major, version.Minor, patch, null, build);
+  }
+
+  /// <summary>
+  /// Gets the current application version.
+  /// </summary>
+  /// <returns>The current application version as <see cref="SemanticVersionObject"/>.</returns>
+  public static SemanticVersionObject GetCurrentAppVersion()
+  {
+    Assembly entryAssembly = Assembly.GetEntryAssembly()
+      ?? throw new InvalidOperationException("Unable to get entry assembly.");
+    AssemblyName assemblyName = entryAssembly.GetName()
+      ?? throw new InvalidOperationException("Unable to get assembly name.");
+
+    if (assemblyName.Version is null)
+    {
+      throw new InvalidOperationException("Unable to get assembly version.");
+    }
+
+    return FromVersion(assemblyName.Version, true);
   }
 
   /// <summary>
@@ -128,7 +285,7 @@ public class SemanticVersionObject : ISemanticVersion, IComparable
   {
     if (obj == null)
     {
-      return 1;
+      return -1;
     }
 
     if (obj is not SemanticVersionObject version)
@@ -167,6 +324,11 @@ public class SemanticVersionObject : ISemanticVersion, IComparable
       result = $"{result}-{PreRelease}";
     }
 
+    if (!string.IsNullOrEmpty(Build))
+    {
+      result = $"{result}+{Build}";
+    }
+
     return result;
   }
 
@@ -188,5 +350,41 @@ public class SemanticVersionObject : ISemanticVersion, IComparable
   public bool IsOlderThan(SemanticVersionObject? compare)
   {
     return CompareTo(compare) == 1;
+  }
+
+  /// <summary>
+  /// Checks if the current version object is a pre-release version or not.
+  /// </summary>
+  /// <returns><see langword="true"/> if this version is a pre-release, else <see langword="false"/>.</returns>
+  [MemberNotNullWhen(true, nameof(PreRelease))]
+  public bool IsPreRelease()
+  {
+    return !string.IsNullOrEmpty(PreRelease);
+  }
+
+  /// <inheritdoc/>
+  public bool Equals(SemanticVersionObject? other)
+  {
+    return CompareTo(other) == 0;
+  }
+
+  /// <inheritdoc/>
+  public override bool Equals(object? obj)
+  {
+    return Equals(obj as SemanticVersionObject);
+  }
+
+  /// <inheritdoc/>
+  public override int GetHashCode()
+  {
+    unchecked
+    {
+      int hashCode = Major;
+      hashCode = (hashCode * 397) ^ Minor;
+      hashCode = (hashCode * 397) ^ Patch;
+      hashCode = (hashCode * 397) ^ (!string.IsNullOrEmpty(PreRelease) ? PreRelease.GetHashCode() : 0);
+
+      return hashCode;
+    }
   }
 }