+++ CHANGELOG.md
@@ -1,3 +1,55 @@
+# [1.5.0](https://github.com/droidsolutions/semantic-version/compare/v1.4.0...v1.5.0) (2025-04-11)
+
+
+### Features
+
+* add compiler hint for IsPreRelease ([6f0f073](https://github.com/droidsolutions/semantic-version/commit/6f0f073e9d2ccfc39d6134a8eafb2cfab2448e36))
+* upgrade to .NET 9 ([4b0c26e](https://github.com/droidsolutions/semantic-version/commit/4b0c26e59e4c8eab517f64d9ecfabdcd3da38d4f))
+
+# [1.4.0](https://github.com/droidsolutions/semantic-version/compare/v1.3.0...v1.4.0) (2024-03-18)
+
+
+### Features
+
+* add static FromVersion method ([89dd5f9](https://github.com/droidsolutions/semantic-version/commit/89dd5f98c87353f82e33c4cc962680fd08b361dc))
+
+# [1.3.0](https://github.com/droidsolutions/semantic-version/compare/v1.2.0...v1.3.0) (2024-01-26)
+
+
+### Features
+
+* add build metadata ([b525a5f](https://github.com/droidsolutions/semantic-version/commit/b525a5fa3f40f37740cfe6490a662f50eda0e2fe))
+* update to .NET 8 ([298a915](https://github.com/droidsolutions/semantic-version/commit/298a915a985daeb426a0fe7543917874d7fa2995))
+
+# [1.2.0](https://github.com/droidsolutions/semantic-version/compare/v1.1.0...v1.2.0) (2023-02-22)
+
+
+### Bug Fixes
+
+* compare null as lower than not null ([f25ffc6](https://github.com/droidsolutions/semantic-version/commit/f25ffc6f769e58201697f4c4cb14caa157005934))
+
+
+### Features
+
+* explicitly support .NET 7 ([6f16711](https://github.com/droidsolutions/semantic-version/commit/6f16711967272bab417660008c81e369062bade6))
+* implement IEquatable ([4d6654c](https://github.com/droidsolutions/semantic-version/commit/4d6654cb7867e82b1a7eac613233b7527615b7cf))
+* implement operators ([bd9dfdc](https://github.com/droidsolutions/semantic-version/commit/bd9dfdcead0eba85ed9406ebf5b7bb6f0afefa7e))
+
+# [1.1.0](https://github.com/droidsolutions/semantic-version/compare/v1.0.0...v1.1.0) (2022-11-10)
+
+
+### Features
+
+* **SemanticVersionObject:** add IsPreRelease method ([5bb79f7](https://github.com/droidsolutions/semantic-version/commit/5bb79f7eb206aa00e61530db7e0c9bba2a5afe22))
+
+# 1.0.0 (2022-01-27)
+
+
+### Features
+
+* add ISemanticVersion and SemanticVersionObject including Comparer ([df8986e](https://github.com/droidsolutions/semantic-version/commit/df8986e91be001acf076054863bb01f510a98b90))
+* initial release ([01dfdeb](https://github.com/droidsolutions/semantic-version/commit/01dfdebc709c5241248ba4ef40c39bad18cafcb9))
+
 # 1.0.0-develop.1 (2022-01-27)